增加 led 驱动

+++ build.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-CURRENT_DIR=`pwd`
+CURRENT_DIR=$(pwd)
 PATH_WORKSPACE_ROOT=$CURRENT_DIR
 
 if [ ! -d "Build" ]; then
@@ -8,7 +8,8 @@ fi
 
 cd Build
 
-if [ "$1" = "clean" ]; then
+# 这里之所以要判断 CMakeCache.txt 是否存在是为了防止进入的不是 CMake 的目录
+if [ "$1" = "clean" ] && [ -f "CMakeCache.txt" ]; then
     yes | rm -rf *
 fi
